+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+ - MockMqttClient now "publishes" messages to an internal class. Added console command for viewing messages.
 
 ## [13.4.0] - 2021-01-14
 ### Added